+++ src/main.ts
@@ -2,6 +2,6 @@ import { Ana } from './ts/index'
 import '../dist/ana.css'
 
 const A = new Ana()
-const a = A.render()
+const a = A.render
 
-a.App()()
+console.log(a)